virt-api: Extract memorydump handler in its own file
